chore: move ReviewTab into its own file
